+++ taskmonitor/core/celery_queues.py
@@ -1,91 +1,29 @@
 """API for working with celery task queue."""
 
 import concurrent.futures
-import datetime as dt
 import functools
 import itertools
 import json
 from collections import defaultdict
-from typing import List, NamedTuple, Optional
+from typing import List
 
 import redis
 
 from django.conf import settings
-from django.core.cache import cache
-from django.utils import timezone
 
 from allianceauth.services.hooks import get_extension_logger
 from app_utils.logging import LoggerAddTag
 
 from taskmonitor import __title__
 from taskmonitor.app_settings import TASKMONITOR_QUEUED_TASKS_CACHE_TIMEOUT
-from taskmonitor.helpers import extract_app_name
+
+from .tasks_cache import QueuedTasksCache, QueuedTaskShort
 
 PRIORITY_SEP = "\x06\x16"
 DEFAULT_PRIORITY_STEPS = range(10)
 QUEUED_TASKS_CACHE_KEY = "queued-tasks-cache-key"
 
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)
-
-
-class QueuedTaskShort(NamedTuple):
-    """DTO for queued tasks, optimized for size."""
-
-    app_name: str
-    id: int
-    name: str
-    priority: int
-
-    @classmethod
-    def from_dict(cls, obj: dict) -> "QueuedTaskShort":
-        """Create QueuedTaskShort from raw task dict."""
-        if "headers" not in obj:
-            raise ValueError("headers missing in obj")
-        headers = obj["headers"]
-        properties = obj["properties"] if "properties" in obj else {}
-        task_name = headers["task"]
-        return cls(
-            app_name=extract_app_name(task_name),
-            id=headers["id"],
-            name=task_name,
-            priority=properties.get("priority"),
-        )
-
-
-class QueuedTaskCacheEntry(NamedTuple):
-    """Wrapper for storing queued tasks in cache with additional data."""
-
-    tasks: List[QueuedTaskShort]
-    created_at: dt.datetime
-
-
-class CacheApi:
-    """API for working with the cache for queued tasks."""
-
-    def get(self) -> Optional[QueuedTaskCacheEntry]:
-        """Retrieve content of cache or None if cache is expired or invalid."""
-        return cache.get(QUEUED_TASKS_CACHE_KEY)
-
-    def set(self, tasks: List[QueuedTaskShort]) -> QueuedTaskCacheEntry:
-        """Store given tasks in cache."""
-        data = QueuedTaskCacheEntry(tasks=tasks, created_at=timezone.now())
-        cache.set(
-            QUEUED_TASKS_CACHE_KEY,
-            data,
-            timeout=TASKMONITOR_QUEUED_TASKS_CACHE_TIMEOUT,
-        )
-        return data
-
-    def clear(self):
-        """Clear the cache."""
-        cache.delete(QUEUED_TASKS_CACHE_KEY)
-
-    def created_at(self) -> Optional[dt.datetime]:
-        """Return date when cache was created or None when cache is invalid."""
-        data = self.get()
-        if data is None:
-            return None
-        return data.created_at
 
 
 def _redis_client():
@@ -119,11 +57,11 @@ def fetch_tasks() -> List[QueuedTaskShort]:
     """Fetch tasks from queues
     and return as ordered list with oldest task in first position.
     """
-    data = local_cache.get()
+    data = tasks_cache.get()
     if data is None:
         logger.debug("Cache is stale. Fetching new tasks from queue.")
         tasks = _fetch_task_from_all_queues()
-        local_cache.set(tasks)
+        tasks_cache.set(tasks)
     else:
         tasks = data.tasks
         logger.debug("Returning tasks from cache.")
@@ -160,7 +98,7 @@ def clear_tasks(queue_name: str = None):
     r = _redis_client()
     for redis_queue_name in _redis_queue_names(queue_name):
         r.delete(redis_queue_name)
-    local_cache.clear()
+    tasks_cache.clear()
 
 
 def add_tasks(queue_name: str, raw_tasks: list):
@@ -179,4 +117,6 @@ def add_tasks(queue_name: str, raw_tasks: list):
     del tasks_by_priority
 
 
-local_cache = CacheApi()
+tasks_cache = QueuedTasksCache(
+    cache_key=QUEUED_TASKS_CACHE_KEY, timeout=TASKMONITOR_QUEUED_TASKS_CACHE_TIMEOUT
+)